requirements.txt
alembic==1.13.2
blinker==1.8.2
cffi==1.17.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
Deprecated==1.2.14
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==23.0.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
limits==3.13.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.2
rich==13.8.1
setuptools==70.3.0
SQLAlchemy==2.0.29
sqlalchemy-serializer==1.4.22
typing_extensions==4.12.2
Werkzeug==3.0.4
wrapt==1.16.0
